ммммм

+++ Aquatic Life-Just Fish, No Chips/Classes/ActiveAquaSitting/IActiveAquaSitting.cs
@@ -9,8 +9,8 @@ namespace Aquatic_Life_Just_Fish__No_Chips.Classes.ActiveAquaSitting
 {
     public interface IActiveAquaSitting : IAquaSitting
     {
-        double Health { get; set; }
-        double Hunger { get; set; }
+        float Health { get; set; }
+        float Hunger { get; set; }
         float VisionRange { get; set; }
         double CurrentAngle { get; set; }
         void UpdateBehavior(AquariumContent content);
